# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.CreatePackageRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      question_ids:
        items:
          type: string
        type: array
      status:
        type: string
    required:
    - code
    - name
    - question_ids
    type: object
  entity.CreateQuestionRequest:
    properties:
      answer_key:
        type: string
      category_id:
        type: string
      code:
        type: string
      difficulty:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.QuestionOptionInput'
        type: array
      question_text:
        type: string
      status:
        type: string
      type:
        type: string
      weight:
        type: number
    required:
    - code
    - question_text
    - type
    type: object
...
  entity.DeletePackageRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
  entity.DeleteQuestionRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
...
  entity.GetPackageListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.PackageListItem'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
  entity.GetQuestionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.QuestionListItem'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  entity.PackageDetailResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      questions:
        items:
          $ref: '#/definitions/entity.PackageQuestionInfo'
        type: array
      status:
        type: string
      updated_at:
        type: string
      verification_status:
        type: string
    type: object
  entity.PackageListItem:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      question_count:
        type: integer
      status:
        type: string
      verification_status:
        type: string
    type: object
  entity.PackageQuestionInfo:
    properties:
      question_code:
        type: string
      question_id:
        type: string
      sort_order:
        type: integer
    type: object
...
  entity.QuestionDetailResponse:
    properties:
      answer_key:
        type: string
      category_id:
        type: string
      category_name:
        type: string
      code:
        type: string
      created_at:
        type: string
      difficulty:
        type: string
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.QuestionOption'
        type: array
      question_text:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
      verification_status:
        type: string
      weight:
        type: number
    type: object
  entity.QuestionListItem:
    properties:
      answer_key:
        type: string
      category_id:
        type: string
      category_name:
        type: string
      code:
        type: string
      created_at:
        type: string
      difficulty:
        type: string
      id:
        type: string
      question_text:
        type: string
      status:
        type: string
      type:
        type: string
      verification_status:
        type: string
      weight:
        type: number
    type: object
  entity.QuestionOption:
    properties:
      id:
        type: string
      is_correct:
        type: boolean
      option_key:
        type: string
      option_text:
        type: string
      question_id:
        type: string
    type: object
  entity.QuestionOptionInput:
    properties:
      is_correct:
        type: boolean
      option_key:
        type: string
      option_text:
        type: string
    type: object
...
  entity.UpdatePackageRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      question_ids:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
  entity.UpdateQuestionRequest:
    properties:
      answer_key:
        type: string
      category_id:
        type: string
      code:
        type: string
      difficulty:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.QuestionOptionInput'
        type: array
      question_text:
        type: string
      status:
        type: string
      type:
        type: string
      weight:
        type: number
    type: object
...
  helper.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend service (hexagon/clean architecture).
  title: Widyaprada Backend API
  version: "1.0"
paths:
...
  /api/v1/exams/{id}/unverify:
    post:
      consumes:
      - application/json
      description: Cancel verification of exam
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unverify exam
      tags:
      - exams
  /api/v1/exams/{id}/verify:
    post:
      consumes:
      - application/json
      description: Mark exam as verified
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify exam
      tags:
      - exams
  /api/v1/question-packages:
    get:
      consumes:
      - application/json
      description: Get paginated list of question packages with filters
      parameters:
      - description: Search
        in: query
        name: q
        type: string
      - description: Draft, Aktif, all
        in: query
        name: status
        type: string
      - description: Belum, Sudah, all
        in: query
        name: status_verifikasi
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetPackageListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List question packages
      tags:
      - question-packages
    post:
      consumes:
      - application/json
      parameters:
      - description: Create package request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreatePackageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.PackageDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create question package
      tags:
      - question-packages
  /api/v1/question-packages/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete package by ID, reason required
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeletePackageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete question package
      tags:
      - question-packages
    get:
      consumes:
      - application/json
      description: Get question package by ID with list of questions
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.PackageDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get package detail
      tags:
      - question-packages
    put:
      consumes:
      - application/json
      description: Update question package by ID
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      - description: Update package request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePackageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.PackageDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update question package
      tags:
      - question-packages
...
  /api/v1/questions:
    get:
      consumes:
      - application/json
      description: Get paginated list of questions with filters
      parameters:
      - description: Search
        in: query
        name: q
        type: string
      - description: Filter by type (PG, BENAR_SALAH, ESSAY)
        in: query
        name: tipe
        type: string
      - description: Filter by category
        in: query
        name: kategori_id
        type: string
      - description: Draft, Aktif, all
        in: query
        name: status
        type: string
      - description: Belum, Sudah, all
        in: query
        name: status_verifikasi
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort by
        in: query
        name: sort_by
        type: string
      - description: asc, desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetQuestionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List questions
      tags:
      - questions
    post:
      consumes:
      - application/json
      description: Create new question (Super Admin only)
      parameters:
      - description: Create question request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.QuestionDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create question
      tags:
      - questions
  /api/v1/questions/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete question by ID, reason required (Super Admin only)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete question
      tags:
      - questions
    get:
      consumes:
      - application/json
      description: Get question by ID with options
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.QuestionDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get question detail
      tags:
      - questions
    put:
      consumes:
      - application/json
      description: Update question by ID (Super Admin only)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Update question request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.QuestionDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update question
      tags:
      - questions
...
swagger: "2.0"